[feat] : BottlesBottomBar 구현
